* commit 'f8a0c455b51a62f6585c00f9aa129d12a19a1f9f':
  Correct int types. Use uint32_t where appropriate

+++ elf_load.c
@@ -59,7 +59,7 @@ static const struct memorymap *get_host_mapping(unsigned long auxcore)
  * A very simple elf loader, assumes the image is valid, returns the
  * entry point address.
  */
-unsigned long load_elf_image_phdr(int fd, char *addr)
+uint32_t load_elf_image_phdr(int fd, char *addr)
 {
     Elf32_Ehdr *ehdr; /* Elf header structure pointer */
     Elf32_Phdr *phdr; /* Program header structure pointer */
